apps/ledger/serializers.py
from .models import LedgerEntry
from rest_framework import serializers

class LedgerEntrySerializer(serializers.ModelSerializer):
    class Meta:
        model = LedgerEntry
        fields = '__all__'