pyproject.toml
[tool.poetry]
name = "demowallet"
version = "0.1.0"
description = ""
authors = ["maxx_PC <m.petrykin@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.1"
django-email-users = "^2.1.0"
drf-yasg = "^1.21.10"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
